+++ main/src/main/java/net/oneandone/stool/extensions/Logstash.java
@@ -18,6 +18,7 @@ package net.oneandone.stool.extensions;
 import net.oneandone.stool.stage.Stage;
 import net.oneandone.stool.util.Files;
 import net.oneandone.sushi.fs.file.FileNode;
+import net.oneandone.sushi.util.Separator;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -57,11 +58,11 @@ public class Logstash implements Extension {
     public void beforeStart(Stage stage) throws IOException {
         String bin;
         FileNode file;
+        StringBuilder config;
 
         bin = link; // TODO: rename
         file = conf(stage);
-        file.writeString(
-                "input {\n" +
+        config = new StringBuilder("input {\n" +
                 "  file {\n" +
                 "    type => 'generic'\n" +
                 "    tags => ['" + stage.getName() + "']\n" +
@@ -69,9 +70,11 @@ public class Logstash implements Extension {
                 "    ignore_older => 0\n" +
                 "    path => ['" + stage.getBackstage().join(Pustefix.APPLOGS).getAbsolute() + "/*/*.log']\n" +
                 "  }\n" +
-                "}\n" +
-                "\n" +
-                output);
+                "}\n\n");
+        for (String name : Separator.COMMA.split(output)) {
+            config.append(file.getWorld().file(name));
+            config.append('\n');
+        }
         Files.stoolFile(file);
         file.getParent().exec(bin, file.getAbsolute(), log(stage).getAbsolute(), pid(stage).getAbsolute());
     }