+++ main/src/main/java/net/oneandone/stool/cli/Main.java
@@ -18,7 +18,6 @@ package net.oneandone.stool.cli;
 import net.oneandone.inline.Cli;
 import net.oneandone.inline.Console;
 import net.oneandone.inline.commands.PackageVersion;
-import net.oneandone.stool.ssl.Pair;
 import net.oneandone.stool.util.LogOutputStream;
 import net.oneandone.stool.util.Logging;
 import net.oneandone.stool.util.Session;
@@ -221,7 +220,7 @@ public class Main {
 
     //--
 
-    private static final List<String> LAZY_HOSTS = Arrays.asList(Pair.HOSTNAME, "api-next.pki.1and1.org");
+    private static final List<String> LAZY_HOSTS = Arrays.asList("api-next.pki.1and1.org");
 
     public static World world() throws IOException {
         World world;