+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ### 3.4.3 (pending)
 
-* fixed stool create gav in shared mode (thanks to Maximilian Braun)
+* fixed `stool create` gav in shared mode (thanks to Maximilian Braun)
 * Dashboard: removed cancel button, it had no effect (thanks to Cosmin Barbu)
 * one logstash process per stage
+* improved logstash
+  * logstash.link now configures a bash script to launch a logstash process 
+  * logstash.output now specifies a list of files to append for the config
 
 
 ### 3.4.2 (2016-08-11)